requirements.prod.txt
Django==4.2.9
djangorestframework==3.14.0
drf-yasg==1.21.7
environs==10.3.0
pillow==10.2.0
unidecode==1.3.8
djoser==2.2.2
django-filter==23.5
django-cors-headers==4.3.1
tqdm==4.66.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
hiredis==2.3.2
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
